Fix line break on FormulaBar
