+++ HISTORY.md
@@ -1,0 +1,21 @@
+# 1.x.x
+
+## 1.0.x
+
+### 1.0.0
+- It is now possible to refer to another sheet. #49
+- Formula bar. #49
+- Add E2E tests. #48
+
+### 1.0.1
+- Bugfix: Cells referenced by formulas do not revert to their original state with undo.
+
+### 1.0.2
+- Bugfix: AutoFill overwrites protected cells.
+
+### 1.0.3, 1.0.4
+- Bugfix: FormulaBar style.
+
+### 1.0.5
+- Bugfix: Autofill does not work when multiple cells are selected. #55, #56
+
