+++ src/components/Editor.tsx
@@ -23,6 +23,7 @@ import { Context } from '../store';
 import { areaToZone } from '../lib/structs';
 import { DEFAULT_HEIGHT } from '../constants';
 import * as prevention from '../lib/prevention';
+import { insertNewLineAtCursor } from '../lib/input';
 
 export const Editor: React.FC = () => {
   const { store, dispatch } = React.useContext(Context);
@@ -102,11 +103,7 @@ export const Editor: React.FC = () => {
       case 'Enter': // ENTER
         if (editing) {
           if (e.altKey) {
-            const selectPoint = input.selectionEnd;
-            const before = input.value.slice(0, selectPoint);
-            const after = input.value.slice(selectPoint);
-            input.value = `${before}\n${after}`;
-            input.selectionEnd = before.length + 1;
+            insertNewLineAtCursor(input);
             input.style.height = `${input.clientHeight + DEFAULT_HEIGHT}px`;
             return false;
           } else {
